+++ src/TsLinkPhp/BaseCL.ts
@@ -2,6 +2,8 @@
 export class BaseCL {
 	public context : any = {};
 
+	public foo = 42;
+
 	public url : string = "";
 
 	public onPrepareRequest : ((ev: ClCallEvent, data : any) => any)|null = null;
@@ -15,20 +17,60 @@ export class BaseCL {
 	// @ts-ignore
 	protected async callMethod(methodName : string, args : any/* : IArguments*/, callOpts : CallOpts = { rawResult: false }, newDataType = null) : Promise<any> {
 
+		const newArgs= [];
+		const uploads: Record<string, File> = {};
+		let c = 0;
+		let i = 0;
+		const uploadArgs = [];
+		for (const arg of args) {
+			if (arg instanceof File) {
+				uploads[c] = arg;
+				newArgs.push(c);
+				c++;
+				uploadArgs.push(i);
+			} else if (arg instanceof FileList) {
+				const newArg: any[] = [];
+				// @ts-ignore
+				for (const file of arg) {
+					uploads[c] = file;
+					newArg.push(c);
+					c++;
+					uploadArgs.push(i);
+				}
+				newArgs.push(newArg);
+			} else {
+				newArgs.push(arg);
+			}
+			i++;
+		}
+
 		const srcData = {
 			name: methodName,
 			context: this.context,
-			pars: [...args]
+			pars: newArgs,
+			uploadArgs,
 		};
+
+		let contentType: string|null;
+		let postData: string|FormData;
+		if (uploadArgs.length > 0) {
+			contentType = null;
+			postData = new FormData();
+			postData.append('request', JSON.stringify(srcData));
+			for (const k in uploads) {
+				postData.append(k, uploads[k]);
+			}
+		} else {
+			contentType = 'application/json';
+			postData = JSON.stringify(srcData);
+		}
 
 		const ev: ClCallEvent = {
 			url: this.url,
 			requestInit: {
-				method: 'POST', // Metoda požadavku
-				body: JSON.stringify(srcData),
-				headers: {
-					'Content-Type': 'application/json',
-				},
+				method: 'POST',
+				body: postData,
+				headers: contentType ? { 'Content-Type': contentType } : {},
 			}
 		}